RESTful cache wraps Local cache
Replaces the CombinedCache approach to composing a remote and local
cache with a remote cache that simply wraps a local one, delegating
to it to store and create tarfiles.

Combined cache can "backfill" caches with artifacts found in other
caches but as it allows composing arbitrary cache implementations,
it must do so via generic insertion calls. In the common case of
remote and local cache, this means downloading a compressed tarball,
extracting it, then recompressing the contents into an identical
tarball, adding almost 20-25% to total cache fetch time.

This removes all the tar file operations from remote cache in favor
of delegating to the embedded local cache, so when a local cache is not
provided to the RESTfulCache constructor, one that provides the same semantics
via tempfiles is used instead.